Fixes for https://github.com/ORNL-QCI/tnqvm/issues/91
It was a copy and paste error when  I tried to fix https://github.com/ORNL-QCI/tnqvm/issues/89

I copied the leg pattern from Hadamard gate which was actually not generally correct (okay for H matrix, which is symmetric).

Effectively, it was applying U gates in reverse.

Need to swap the leg order.

Tested by: running both the Python test scripts.

Signed-off-by: Thien Nguyen <nguyentm@ornl.gov>